requirements.txt
pandas
torch
faiss-cpu
sentence-transformers
transformers
gradio